+++ lifelines/fitters/kaplan_meier_fitter.py
@@ -6,7 +6,7 @@ import pandas as pd
 
 from lifelines.fitters import UnivariateFitter
 from lifelines.utils import _preprocess_inputs, _additive_estimate, StatError, inv_normal_cdf,\
-    median_survival_times
+    median_survival_times, check_nans
 from lifelines.plotting import plot_loglogs
 
 
@@ -22,7 +22,7 @@ class KaplanMeierFitter(UnivariateFitter):
     """
 
     def fit(self, durations, event_observed=None, timeline=None, entry=None, label='KM_estimate',
-            alpha=None, left_censorship=False, ci_labels=None):
+            alpha=None, left_censorship=False, ci_labels=None, weights=None):
         """
         Parameters:
           duration: an array, or pd.Series, of length n -- duration subject was observed for
@@ -38,16 +38,24 @@ class KaplanMeierFitter(UnivariateFitter):
           left_censorship: True if durations and event_observed refer to left censorship events. Default False
           ci_labels: add custom column names to the generated confidence intervals
                 as a length-2 list: [<lower-bound name>, <upper-bound name>]. Default: <label>_lower_<alpha>
-
+          weights: n array, or pd.Series, of length n, if providing a weighted dataset. For example, instead
+              of providing every subject as a single element of `durations` and `event_observed`, one could
+              weigh subject differently.
 
         Returns:
           self, with new properties like 'survival_function_'.
 
         """
+
+        check_nans(durations)
+        if event_observed is not None:
+          check_nans(event_observed)
+
         # if the user is interested in left-censorship, we return the cumulative_density_, no survival_function_,
         estimate_name = 'survival_function_' if not left_censorship else 'cumulative_density_'
-        v = _preprocess_inputs(durations, event_observed, timeline, entry)
+        v = _preprocess_inputs(durations, event_observed, timeline, entry, weights)
         self.durations, self.event_observed, self.timeline, self.entry, self.event_table = v
+
         self._label = label
         alpha = alpha if alpha else self.alpha
         log_survival_function, cumulative_sq_ = _additive_estimate(self.event_table, self.timeline,